Initial attempt of moving logic out of console
